refactor: 외부 라이브러리 의존성 낮추도록 리팩토링
