+++ src/main/java/nextstep/subway/line/domain/Lines.java
@@ -1,6 +1,5 @@
 package nextstep.subway.line.domain;
 
-import nextstep.subway.station.domain.Station;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
@@ -15,12 +14,12 @@ public class Lines {
     }
 
     public DijkstraShortestPath createPath() {
-        WeightedMultigraph<Station, DefaultWeightedEdge> graph = new WeightedMultigraph(DefaultWeightedEdge.class);
+        SectionGraph graph = new SectionGraph(new WeightedMultigraph(DefaultWeightedEdge.class));
 
         for (Line line : lines) {
             line.registerPath(graph);
         }
 
-        return new DijkstraShortestPath(graph);
+        return graph.getPath();
     }
 }
